Switch format to 32 bit floating point @ 44100 Hz

+++ sound.h
@@ -1,6 +1,4 @@
-#include <stdint.h>
-
-typedef uint8_t songbuf_t;
+typedef float songbuf_t;
 
 int sound_open();
 int sound_play(songbuf_t *buf, size_t bufsize);
